Feat/dashboard (#4)
* fix(todo-item): corrected interaction with local-storage service

* refactor(todo-item): extended "save" method in repository

+++ src/todo-item/service/todo-item.service.ts
@@ -6,19 +6,30 @@ import type {ITodoItemRepository} from '../repository';
 import type {ITodoItemCreateFormDto} from '../components';
 
 class TodoItemService implements ITodoItemService {
-  constructor(private readonly _repo: ITodoItemRepository) {
-    makeAutoObservable(this);
-  }
-
   private _todos: Array<TodoItemModel> = [];
 
   get todos(): ReadonlyArray<TodoItemModel> {
     return this._todos;
   }
 
+  constructor(private readonly _repo: ITodoItemRepository) {
+    makeAutoObservable(this);
+  }
+
+  async init(todos: TodoItemModel[]): Promise<void> {
+    const storageTodos = await this._repo.list().catch(console.error);
+
+    if (!storageTodos?.length) {
+      this._todos = todos;
+      await Promise.all(this._todos.map((item) => this._repo.save(item))).catch(console.error);
+    } else {
+      this._todos = storageTodos;
+    }
+  }
+
   async addTodo(dto: ITodoItemCreateFormDto): Promise<void> {
     const todo = TodoItemFactory.fromFormCreateDto(dto);
-    this._todos = [todo, ...this._todos];
+    this._todos = [...this._todos, todo];
 
     await this._repo.save(todo).catch(console.error);
   }
@@ -26,9 +37,9 @@ class TodoItemService implements ITodoItemService {
   async deleteTodo(todoId: string): Promise<void> {
     const idx = this._todos.findIndex((todo) => todo.id === todoId);
 
-    if (idx) {
-      this._todos.splice(idx, 1);
+    if (isFinite(idx) && idx !== -1) {
       const todo = this._todos[idx];
+      this._todos.splice(idx, 1);
       await this._repo.remove(todo).catch(console.error);
     }
   }
@@ -36,7 +47,7 @@ class TodoItemService implements ITodoItemService {
   async editTodo(todoId: string, dto: ITodoItemCreateFormDto): Promise<void> {
     const idx = this.todos.findIndex((item) => item.id === todoId);
 
-    if (idx) {
+    if (isFinite(idx) && idx !== -1) {
       const todo = this._todos[idx];
       const newTodo = TodoItemFactory.fromFormEditDto(dto, todo);
       this._todos.splice(idx, 1, newTodo);
@@ -45,20 +56,14 @@ class TodoItemService implements ITodoItemService {
   }
 
   async toggleTodo(todoId: string): Promise<void> {
-    const idx = this.todos.findIndex((item) => item.id === todoId);
+    const idx = this._todos.findIndex((item) => item.id === todoId);
 
-    if (idx) {
+    if (isFinite(idx) && idx !== -1) {
       const todo = this._todos[idx];
       todo.completed = !todo.completed;
       this._todos.splice(idx, 1, todo);
       await this._repo.save(todo).catch(console.error);
     }
-  }
-
-  async init(todos: TodoItemModel[]): Promise<void> {
-    this._todos = todos;
-
-    await Promise.all(this._todos.map((item) => this._repo.save(item))).catch(console.error);
   }
 }
